requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0
django-appconf==1.0.5
django-celery-beat==2.1.0
django-crispy-forms==1.14.0
django-excel==0.0.10
django-import-export==2.8.0
django-timeline-logger==2.1.0
django-timezone-field==4.2.3
et-xmlfile==1.1.0
idna==3.3
iso8601==0.1.16
jdcal==1.4.1
kombu==5.2.4
lml==0.1.0
MarkupPy==1.14
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
pkg_resources==0.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
rapidpro-python==2.8.5
redis==4.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
texttable==1.6.4
unicodecsv==0.14.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
